Create a markdown table from prepreprint summaries

Create a markdown table from prepreprint summaries

Usage

tt_preprints(preprints, cols = c("title", "summary"), width = c(1, 3))

Arguments

preprints

Output from get_preprints() followed by add_prompt() followed by add_summary().

cols

Columns to display in the resulting markdown table.

width

Vector of relative widths equal to length(cols).

Value

A tinytable table.

Examples

# Use built-in example data
example_preprints
#> # A tibble: 60 × 7
#>    source  subject        title                    url   abstract prompt summary
#>    <chr>   <chr>          <chr>                    <chr> <chr>    <chr>  <chr>  
#>  1 bioRxiv bioinformatics MedGraphNet: Leveraging… http… Genetic… I am … MedGra…
#>  2 bioRxiv bioinformatics High-throughput bacteri… http… The com… I am … The co…
#>  3 bioRxiv bioinformatics scParadise: Tunable hig… http… scRNA-s… I am … scAdam…
#>  4 bioRxiv bioinformatics Camera Paths, Modeling,… http… The enh… I am … ArtiaX…
#>  5 bioRxiv bioinformatics dScaff - an automatic b… http… Rapid e… I am … dScaff…
#>  6 bioRxiv bioinformatics Jaeger: an accurate and… http… Abstrac… I am … Jaeger…
#>  7 bioRxiv bioinformatics AI-Augmented R-Group Ex… http… Efficie… I am … The pa…
#>  8 bioRxiv bioinformatics OPLS-based Multiclass C… http… Multicl… I am … OPLS-D…
#>  9 bioRxiv bioinformatics STANCE: a unified stati… http… A signi… I am … STANCE…
#> 10 bioRxiv bioinformatics AsaruSim: a single-cell… http… Motivat… I am … AsaruS…
#> # ℹ 50 more rows
tt_preprints(example_preprints[1:2,])
#> 
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
#> | title                                                                                                                                                                          | summary                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            |
#> +================================================================================================================================================================================+====================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================+
#> | [MedGraphNet: Leveraging Multi-Relational Graph Neural Networks and Text Knowledge for Biomedical Predictions](http://biorxiv.org/cgi/content/short/2024.09.24.614782v1?rss=1) | MedGraphNet leverages multi-relational Graph Neural Networks and text knowledge to improve biomedical predictions by initializing nodes using informative embeddings from existing text knowledge, allowing for robust integration of various data types and improved generalizability. The model demonstrates superior performance compared to traditional single-relation approaches in scenarios with isolated or sparsely connected nodes, particularly in identifying disease-gene associations and drug-phenotype relationships, and shows promising results in accurately inferring drug side effects without direct training on such data. |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
#> | [High-throughput bacterial aggregation analysis in droplets](http://biorxiv.org/cgi/content/short/2024.09.24.613170v1?rss=1)                                                   | The communal lifestyle of bacteria can contribute significantly to antimicrobial resistance by promoting biofilm formation. A key approach to addressing this issue is to develop novel techniques for analyzing bacterial behavior, such as those enabled by droplet-based platforms and image analysis methods.                                                                                                                                                                                                                                                                                                                                  |
#> +--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+